refactor: type safe
