+++ src/app/api/twitter/userTimeline/[userId]/liked/route.ts
@@ -8,7 +8,7 @@ import { createTwitterClientV2 } from '@/helpers/createTwitterClientV2.js';
 import { getSearchParamsFromRequestWithZodObject } from '@/helpers/getSearchParamsFromRequestWithZodObject.js';
 import { withRequestErrorHandler } from '@/helpers/withRequestErrorHandler.js';
 import { withTwitterRequestErrorHandler } from '@/helpers/withTwitterRequestErrorHandler.js';
-import { Pageable } from '@/schemas/Pageable.js';
+import { Pageable } from '@/schemas/index.js';
 import type { NextRequestContext } from '@/types/index.js';
 
 export const GET = compose<(request: NextRequest, context?: NextRequestContext) => Promise<Response>>(
